Refactor code for creating slugs and global problems
The changes include refactoring parts of the code responsible for slug creation and global problem creation into separate helper functions for better reusability and code maintainability. Additionally, certain import statements have been modified for better organization and some redundant parts of the code were removed for cleaner codebase.
