+++ lib/stringHelper.ts
@@ -1,0 +1,6 @@
+export function createSlug(input: string): string {
+    return input
+        .toLowerCase()
+        .replace(/[^a-z0-9]+/g, '-')
+        .replace(/^-+|-+$/g, '');
+}
